skip sudo file if it exists

+++ gs-install.sh
@@ -199,10 +199,12 @@ else
     echo -e "${INFO} Executing Gravity Sync Deployment"
     
     if [ "$LOCALADMIN" == "sudo" ]; then
-        echo -e "${STAT} Creating sudoers.d permissions file"
-        touch /tmp/gs-nopasswd.sudo
-        echo -e "${CURRENTUSER} ALL=(ALL) NOPASSWD: ALL" > /tmp/gs-nopasswd.sudo
-        sudo install -m 0440 /tmp/gs-nopasswd.sudo /etc/sudoers.d/gs-nopasswd
+        if [ ! -f /etc/sudoers.d/gs-nopasswd ]; then
+            echo -e "${STAT} Creating sudoers.d permissions file"
+            touch /tmp/gs-nopasswd.sudo
+            echo -e "${CURRENTUSER} ALL=(ALL) NOPASSWD: ALL" > /tmp/gs-nopasswd.sudo
+            sudo install -m 0440 /tmp/gs-nopasswd.sudo /etc/sudoers.d/gs-nopasswd
+        fi
     fi
 
     if [ -f /etc/bash.bashrc ]; then
